requirements.txt
Django==5.0.2
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
python-dotenv==1.0.1
django-environ==0.11.2
django-compressor==4.4
django-storages==1.14.2
django-allauth==0.61.1
django-ckeditor==6.7.1
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-celery-beat==2.6.0
django-celery-results==2.5.1
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
rcssmin==1.1.1
rjsmin==1.2.1
sentry-sdk==1.40.6